# docs/b2b-core/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PharmAssist B2B Core API (/v1)",
    "version": "0.1.0",
    "description": "The B2B /v1 surface for integrators: Tier-1 Core, the `clinical_only` base tier below it, and the Tier-2 routes gated at `clinical` (a tier-gated route documents its 403). Authenticate with the `X-API-Key` header over TLS. Errors render in the stable `{\"error\": {code, message, request_id}}` envelope. See docs/b2b-core/API.md for the narrative (auth, tiers, locations without ΗΔΥΚΑ credentials, error codes, consent attestation, the 609/ΕΟΠΥΥ rule, rate limits, tri-state coverage, pagination, sandbox identifiers)."
  },
  "servers": [
    {
      "url": "https://api.<your-domain>",
      "description": "Live (FT-10, TBD)"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/v1/adr-reports": {
      "post": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Create Adr Report",
        "description": "Create a new ADR report for this location.  Status starts at PENDING_REVIEW.\n\nEOF_REPORTED is reachable via /transition; PharmAssist does NOT submit to ΕΟΦ.",
        "operationId": "create_adr_report_v1_adr_reports_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdrReportCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `clinical` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "List Adr Reports",
        "description": "List ADR reports for this location with optional filters.\n\nFilters: amka, atc (ATC code), barcode (medicine barcode), status, from (date), to (date).\nSorted by reportedAt descending. Paginated.",
        "operationId": "list_adr_reports_v1_adr_reports_get",
        "parameters": [
          {
            "name": "amka",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Amka"
            }
          },
          {
            "name": "atc",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by ATC code",
              "title": "Atc"
            },
            "description": "Filter by ATC code"
          },
          {
            "name": "barcode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by medicine barcode",
              "title": "Barcode"
            },
            "description": "Filter by medicine barcode"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `clinical` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/adr-reports/{report_id}": {
      "get": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Get Adr Report",
        "description": "Fetch a single ADR report (with event history).  404 if not owned by this location.",
        "operationId": "get_adr_report_v1_adr_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `clinical` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/adr-reports/{report_id}/transition": {
      "post": {
        "tags": [
          "b2b-v1",
          "b2b-v1"
        ],
        "summary": "Transition Adr Report",
        "description": "Advance the ADR report to the next status.\n\nTransitions: PENDING_REVIEW → ESCALATED → EOF_REPORTED → CLOSED.\nCLOSED is terminal — transitioning from it returns 409 illegal_transition.\nEOF_REPORTED means the report is marked; PharmAssist does NOT submit to ΕΟΦ.",
        "operationId": "transition_adr_report_v1_adr_reports__report_id__transition_post",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdrTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "`tier_required` — this endpoint requires the `clinical` tier or higher; the customer's entitlement tier ranks below it."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/status": {
      "get": {
        "tags": [
          "b2b-v1"
        ],
        "summary": "V1 Status",
        "description": "Auth smoke endpoint: who am I, and is my location's ΗΔΥΚΑ session warm.\n\n`location.retrievalAvailable: false` means the location was provisioned\nwithout ΗΔΥΚΑ credentials — patient/prescription routes answer 409\n`retrieval_unavailable`; drugs and safety work. `pharmapiConnected: false`\nis normal before the first upstream call — sessions are established lazily\nper location — and always false when retrieval is unavailable.",
        "operationId": "v1_status_v1_status_get",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdrReportCreate": {
        "properties": {
          "patientAmka": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patientamka"
          },
          "patientName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patientname"
          },
          "rxId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rxid"
          },
          "medicineBarcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicinebarcode"
          },
          "medicineName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medicinename"
          },
          "atcCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atccode"
          },
          "symptomDescription": {
            "type": "string",
            "title": "Symptomdescription"
          },
          "onsetTiming": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onsettiming"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "causality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causality"
          },
          "eofReportRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eofreportref"
          }
        },
        "type": "object",
        "required": [
          "symptomDescription"
        ],
        "title": "AdrReportCreate"
      },
      "AdrTransitionRequest": {
        "properties": {
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "AdrTransitionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Per-location bearer key (`pa_live_…` / `pa_test_…`). TLS only. All failure modes return one indistinguishable 401."
      }
    }
  }
}